Changed weight and weight_achieved to floats

+++ Task.h
@@ -4,14 +4,14 @@ class Task{
 
     private:
 
-        unsigned int weight;
-        unsigned int weight_achieved;
+        float weight;
+        float weight_achieved;
         float grade;
 
     public:
 
         std::string name;
         std::string description;
-        Task(std::string name, std::string description, int weight );
+        Task(std::string name, std::string description, float weight );
 
 };
